Fixing Faked Partitions, still double counting

+++ db-engine/src/main/java/com/splicemachine/db/iapi/stats/TableStatisticsImpl.java
@@ -154,7 +154,7 @@ public class TableStatisticsImpl implements TableStatistics {
             selectivity =+ partitionStatistic.rangeSelectivity(start,stop,includeStart,includeStop,positionNumber);
             rowCount =+ partitionStatistic.rowCount();
         }
-        return selectivity/rowCount;
+        return (double) ((double)selectivity/(double)rowCount);
     }
 
     @Override
